refactor(meta-governor): split core type contracts
